Successfully created sparse GeoTIFF.
$ python cliptest.py
Opening data/ucl_elie/ESACCI-LC-L4-LCCS-Map-300m-P1Y-2015-v2.0.7.tif
opening shapefile
Creating new shapefile
Rasterizing

$ ls -l result.tif
-rw-r--r--  1 dgentry  staff  1444833 Sep 28 15:20 result.tif

$ python cliptest.py
full: 6685  hole: 122093 partial: 0 error: 0

Note: 2387562 bytes for LZW of a non-sparse file, as the empty
areas compress really well. Omitting 95% of the blocks reduces
the file size by 40%. However, the main reason we're doing this
is not file size, it is so the consumer of these masks can skip
processing 95% of the blocks altogether.

Note2: As advertised in the documentation, other TIFF-supporting
software like MacOS Preview cannot grok a sparse file like this.
It says the file has been damaged. We may make a deliberately
different filename extension to make this not be so bad.