Checkpoint perftest before removal.
Improving the performance of counting pixels was interesting
but not ultimately useful as cutline processing to extract
countries from the image turned out to be the bulk of the time
spent. GDAL does almost everything with SWIG native code, which
made cProfile difficult to figure out and led to working on the
wrong thing.