Checkpoint rewrite classify_geomorpho90m_slope.py
New version works on the 90m files, and pre-processes them into
a 1km GeoTIFF which matches the dimensions of the data/Beck_KG_V1
0p0083 files.

The classified file has 8 bands, one per slope class, giving the
percentage as an integer. As there are exactly 100 pixels in the
original geomorpho90m file for each pixel in the classified file,
integer percentages work well.