Retire the shapefile implementation.
Retire the original implementation which applied the shapefile
to pull features out one by one. It worked well, but was quite slow
due mainly to the expense of processing the shapefile. We
implemented a new technique which preprocesses the shapefile into
rasters of the appropriate resolution in masks/*._tif (+), which
is dramatically faster. All of the data sources now use the masks,
so delete the original code.

Redo the unit tests to use masks.

+ The mask files are named *._tif because they use GDAL's SPARSE_OK
feature to omit empty areas, and they use ZSTD compression, both
of which are not widely supported in TIF utilities like the MacOS
Preview.app. The file extension *._tif is intended to convey that
the file is TIFF-inspired but won't try to make Preview.app open it.